one line updated

+++ study_bud/base/views.py
@@ -4,6 +4,7 @@ from django.contrib import messages
 from django.http import HttpResponse
 from django.shortcuts import redirect, render
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
 from .forms import RoomForm
@@ -35,6 +36,10 @@ def loginPage(request):
     context ={}
     return render (request,'base/login_register.html',context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 def home (request):
     q=request.GET.get('q') if request.GET.get('q')!=None else ''
     #Implementing The Dynamic Searching Functionality
@@ -54,7 +59,9 @@ def room(request,pk):
     context = {'room':room}
  
     return render
-    
+
+
+@login_required(login_url='login')
 def CreateRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -65,11 +72,13 @@ def CreateRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def updateRoom (request,pk):
     room = Room.objects.get(id=pk)
-
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -81,8 +90,12 @@ def updateRoom (request,pk):
     
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         print("Deletion succesful")
